perf: fix blank first screen and shrink critical-path bundles
- add pre-paint theme script and pure-CSS splash screen to index.html
- add root Suspense boundary so i18n loading no longer blocks all commits
- preload locale JSON files in parallel with the JS bundle
- start the global-config request at module scope, parallel with i18n
- lazy-load the markdown stack (416KB gz) and dotlottie player (60KB gz)
- replace mathjs (660KB min) with plain arithmetic in unit-conversions
- split eager tanstack/i18next vendors into cacheable chunks
