+++ apps/admin/index.html
@@ -27,9 +27,68 @@
     <meta name="keywords" content="">
 
     <title>Loading...</title>
+
+    <!-- Apply the saved theme before first paint to avoid a white flash in dark mode -->
+    <script>
+    "use strict";
+    (() => {
+      try {
+        const match = document.cookie.match(/(?:^|;\s*)theme=([^;]*)/);
+        const theme = match ? decodeURIComponent(match[1]) : "system";
+        const dark =
+          theme === "dark" ||
+          (theme !== "light" &&
+            window.matchMedia("(prefers-color-scheme: dark)").matches);
+        if (dark) document.documentElement.classList.add("dark");
+        document.documentElement.style.colorScheme = dark ? "dark" : "light";
+      } catch {
+        /* noop */
+      }
+    })();
+    </script>
+
+    <!-- Splash screen shown until React commits its first content into #app -->
+    <style>
+    html {
+      background-color: oklch(1 0 0);
+    }
+    html.dark {
+      background-color: oklch(0.141 0.005 285.823);
+    }
+    #splash {
+      position: fixed;
+      inset: 0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+    #app:not(:empty) ~ #splash {
+      display: none;
+    }
+    #splash .spinner {
+      width: 2.5rem;
+      height: 2.5rem;
+      border: 3px solid oklch(0.92 0.004 286.32);
+      border-top-color: oklch(0.61 0.11 222);
+      border-radius: 50%;
+      animation: splash-spin 0.8s linear infinite;
+    }
+    html.dark #splash .spinner {
+      border-color: oklch(1 0 0 / 10%);
+      border-top-color: oklch(0.71 0.13 215);
+    }
+    @keyframes splash-spin {
+      to {
+        transform: rotate(360deg);
+      }
+    }
+    </style>
   </head>
   <body>
     <div id="app"></div>
+    <div id="splash">
+      <div class="spinner"></div>
+    </div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
