+++ apps/admin/src/routes/__root.tsx
@@ -5,11 +5,11 @@ import { Toaster } from "@workspace/ui/components/sonner";
 import { NavigationProgress } from "@workspace/ui/composed/navigation-progress";
 import { TanStackQueryDevtools } from "@workspace/ui/integrations/tanstack-query-devtools";
 import { getCookie } from "@workspace/ui/lib/cookies";
-import { getGlobalConfig } from "@workspace/ui/services/common/common";
 import { isBrowser } from "@workspace/ui/utils/index";
 import { useEffect } from "react";
 import { Helmet, HelmetProvider } from "react-helmet-async";
 import { useGlobalStore } from "@/stores/global";
+import { fetchInitialConfig } from "@/utils/bootstrap";
 
 export const Route = createRootRouteWithContext()({
   component: () => {
@@ -17,7 +17,7 @@ export const Route = createRootRouteWithContext()({
     useEffect(() => {
       const initializeApp = async () => {
         try {
-          const configResponse = await getGlobalConfig();
+          const configResponse = await fetchInitialConfig();
           if (configResponse.data?.data) {
             setCommon(configResponse.data.data);
           }
