+++ apps/user/src/main.tsx
@@ -7,7 +7,7 @@ import {
   TanStackQueryContext,
   TanStackQueryProvider,
 } from "@workspace/ui/integrations/tanstack-query";
-import { StrictMode } from "react";
+import { StrictMode, Suspense } from "react";
 import ReactDOM from "react-dom/client";
 // Import the generated route tree
 import { routeTree } from "./routeTree.gen";
@@ -17,30 +17,22 @@ import { DirectionProvider } from "@workspace/ui/integrations/direction";
 import { LanguageProvider } from "@workspace/ui/integrations/language";
 import { ThemeProvider } from "@workspace/ui/integrations/theme";
 import { initializeI18n } from "@workspace/ui/lib/i18n";
-import { fallbackLng, supportedLngs } from "./config/index.ts";
+import { fallbackLng, i18nNamespaces, supportedLngs } from "./config/index.ts";
 // Report web vitals
 import reportWebVitals from "./reportWebVitals.ts";
+import { fetchInitialConfig } from "./utils/bootstrap.ts";
 // Common utilities
 import { Logout } from "./utils/common.ts";
 
 initializeI18n({
   supportedLngs,
   fallbackLng,
-  ns: [
-    "auth",
-    "components",
-    "dashboard",
-    "document",
-    "layout",
-    "main",
-    "order",
-    "payment",
-    "profile",
-    "subscribe",
-    "ticket",
-    "wallet",
-  ],
+  ns: i18nNamespaces,
 });
+
+// Start the global-config request now so it runs in parallel with translation
+// loading; the root route awaits the same promise.
+fetchInitialConfig();
 
 window.logout = Logout;
 
@@ -72,15 +64,19 @@ if (rootElement && !rootElement.innerHTML) {
   const root = ReactDOM.createRoot(rootElement);
   root.render(
     <StrictMode>
-      <TanStackQueryProvider {...TanStackQueryProviderContext}>
-        <LanguageProvider supportedLanguages={supportedLngs}>
-          <ThemeProvider>
-            <DirectionProvider>
-              <RouterProvider router={router} />
-            </DirectionProvider>
-          </ThemeProvider>
-        </LanguageProvider>
-      </TanStackQueryProvider>
+      {/* i18n suspends until translations load; a null fallback keeps #app
+          empty so the index.html splash stays visible until first commit */}
+      <Suspense fallback={null}>
+        <TanStackQueryProvider {...TanStackQueryProviderContext}>
+          <LanguageProvider supportedLanguages={supportedLngs}>
+            <ThemeProvider>
+              <DirectionProvider>
+                <RouterProvider router={router} />
+              </DirectionProvider>
+            </ThemeProvider>
+          </LanguageProvider>
+        </TanStackQueryProvider>
+      </Suspense>
     </StrictMode>
   );
 }
