+++ apps/user/vite.config.ts
@@ -5,6 +5,46 @@ import { devtools } from "@tanstack/devtools-vite";
 import { tanstackRouter } from "@tanstack/router-plugin/vite";
 import viteReact from "@vitejs/plugin-react";
 import { defineConfig, loadEnv, type Plugin } from "vite";
+import { fallbackLng, i18nNamespaces, supportedLngs } from "./src/config/i18n";
+
+// Preload translation files in parallel with the JS bundle. i18next only
+// starts fetching them after the entry chunk has executed, and first render
+// blocks on them, so warming the fetch cache from index.html removes a full
+// network waterfall from the critical path.
+function localePreloadPlugin(): Plugin {
+  const script = `(() => {
+  try {
+    const supported = ${JSON.stringify(supportedLngs)};
+    const fallback = ${JSON.stringify(fallbackLng)};
+    const namespaces = ${JSON.stringify(i18nNamespaces)};
+    let lng = localStorage.getItem("language");
+    if (!(lng && supported.includes(lng))) {
+      const candidates = navigator.languages || [navigator.language];
+      lng = candidates.find((l) => supported.includes(l)) || fallback;
+    }
+    const langs = lng === fallback ? [lng] : [lng, fallback];
+    for (const lang of langs) {
+      for (const ns of namespaces) {
+        const link = document.createElement("link");
+        link.rel = "preload";
+        link.as = "fetch";
+        link.crossOrigin = "anonymous";
+        link.href = "assets/locales/" + lang + "/" + ns + ".json";
+        document.head.appendChild(link);
+      }
+    }
+  } catch {
+    /* noop */
+  }
+})();`;
+
+  return {
+    name: "locale-preload",
+    transformIndexHtml() {
+      return [{ tag: "script", injectTo: "head", children: script }];
+    },
+  };
+}
 
 // Plugin to generate version.lock file after build
 function versionLockPlugin(): Plugin {
@@ -37,6 +77,7 @@ export default defineConfig(({ mode }) => {
       }),
       viteReact(),
       tailwindcss(),
+      localePreloadPlugin(),
       versionLockPlugin(),
     ],
     resolve: {
@@ -55,6 +96,28 @@ export default defineConfig(({ mode }) => {
     },
     build: {
       assetsDir: "static",
+      rollupOptions: {
+        output: {
+          // Keep the always-loaded vendors in their own chunks so app-code
+          // releases don't invalidate them in the browser cache. Only list
+          // packages the entry imports eagerly — grouping lazily-used
+          // @tanstack packages (table, form, ...) here would bloat first load.
+          manualChunks(id: string) {
+            if (!id.includes("node_modules")) return;
+            if (
+              /node_modules\/@tanstack\/(?:react-router|router-[^/]+|history|react-query|query-[^/]+|react-store|store)\//.test(
+                id
+              )
+            ) {
+              return "tanstack";
+            }
+            if (/node_modules\/(?:i18next[^/]*|react-i18next)\//.test(id)) {
+              return "i18n";
+            }
+            return;
+          },
+        },
+      },
     },
   };
 });
