+++ packages/ui/package.json
@@ -15,6 +15,7 @@
     "@dnd-kit/utilities": "^3.2.2",
     "@hookform/resolvers": "^5.2.2",
     "@iconify/react": "^6.0.2",
+    "@lottiefiles/dotlottie-react": "^0.17.7",
     "@monaco-editor/react": "^4.7.0",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-alert-dialog": "^1.1.15",
@@ -56,7 +57,6 @@
     "i18next-http-backend": "^3.0.2",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.554.0",
-    "mathjs": "^15.1.0",
     "monaco-editor": "^0.55.1",
     "monaco-themes": "^0.4.7",
     "motion": "^12.23.24",
